Refactor utility functions into modular files
Moved utility functions from `utils/__init__.py` into separate modules (`matrices.py`, `validation.py`, `device.py`, and `serialization.py`) for improved code organization and maintainability. Updated imports in `__init__.py` to reflect the new structure.
