+++ pytorch_tabnet/utils/matrices.py
@@ -1,0 +1,161 @@
+"""Matrix utility functions for TabNet."""
+
+from typing import List, Union
+
+import numpy as np
+import scipy
+import torch
+
+# def create_explain_matrix(
+#     input_dim: int,
+#     cat_emb_dim: Union[int, List[int]],
+#     cat_idxs: List[int],
+#     post_embed_dim: int,
+# ) -> scipy.sparse.csc_matrix:
+#
+
+
+# def create_group_matrix(list_groups: List[List[int]], input_dim: int) -> torch.Tensor:
+
+
+# def check_list_groups(list_groups: List[List[int]], input_dim: int) -> None:
+def create_explain_matrix(
+    input_dim: int,
+    cat_emb_dim: Union[int, List[int]],
+    cat_idxs: List[int],
+    post_embed_dim: int,
+) -> scipy.sparse.csc_matrix:
+    """Create a reducing matrix for summing importances from embeddings to initial indices.
+
+    In order to rapidly sum importances from same embeddings to the initial index.
+
+    Parameters
+    ----------
+    input_dim : int
+        Initial input dim
+    cat_emb_dim : int or list of int
+        if int : size of embedding for all categorical feature
+        if list of int : size of embedding for each categorical feature
+    cat_idxs : list of int
+        Initial position of categorical features
+    post_embed_dim : int
+        Post embedding inputs dimension
+
+    Returns
+    -------
+    reducing_matrix : np.array
+        Matrix of dim (post_embed_dim, input_dim) to perform reduce
+
+    """
+    if isinstance(cat_emb_dim, int):
+        all_emb_impact = [cat_emb_dim - 1] * len(cat_idxs)
+    else:
+        all_emb_impact = [emb_dim - 1 for emb_dim in cat_emb_dim]
+
+    acc_emb = 0
+    nb_emb = 0
+    indices_trick = []
+    for i in range(input_dim):
+        if i not in cat_idxs:
+            indices_trick.append([i + acc_emb])
+        else:
+            indices_trick.append(
+                range(i + acc_emb, i + acc_emb + all_emb_impact[nb_emb] + 1)  # type: ignore
+            )
+            acc_emb += all_emb_impact[nb_emb]
+            nb_emb += 1
+
+    reducing_matrix = np.zeros((post_embed_dim, input_dim))
+    for i, cols in enumerate(indices_trick):
+        reducing_matrix[cols, i] = 1
+
+    return scipy.sparse.csc_matrix(reducing_matrix)
+
+
+def create_group_matrix(list_groups: List[List[int]], input_dim: int) -> torch.Tensor:
+    """Create the group matrix corresponding to the given list_groups.
+
+    Parameters
+    ----------
+    list_groups : list of list of int
+        Each element is a list representing features in the same group.
+        One feature should appear in maximum one group.
+        Feature that don't get assigned a group will be in their own group of one feature.
+    input_dim : int
+        Number of features in the initial dataset.
+
+    Returns
+    -------
+    group_matrix : torch.Tensor
+        A matrix of size (n_groups, input_dim) where m_ij represents the importance of feature j in group i.
+        The rows must sum to 1 as each group is equally important a priori.
+
+    """
+    check_list_groups(list_groups, input_dim)
+
+    if len(list_groups) == 0:
+        group_matrix = torch.eye(input_dim)
+        return group_matrix
+    else:
+        n_groups = input_dim - int(np.sum([len(gp) - 1 for gp in list_groups]))
+        group_matrix = torch.zeros((n_groups, input_dim))
+
+        remaining_features = [feat_idx for feat_idx in range(input_dim)]
+
+        current_group_idx = 0
+        for group in list_groups:
+            group_size = len(group)
+            for elem_idx in group:
+                # add importrance of element in group matrix and corresponding group
+                group_matrix[current_group_idx, elem_idx] = 1 / group_size
+                # remove features from list of features
+                remaining_features.remove(elem_idx)
+            # move to next group
+            current_group_idx += 1
+        # features not mentionned in list_groups get assigned their own group of singleton
+        for remaining_feat_idx in remaining_features:
+            group_matrix[current_group_idx, remaining_feat_idx] = 1
+            current_group_idx += 1
+        return group_matrix
+
+
+def check_list_groups(list_groups: List[List[int]], input_dim: int) -> None:
+    """Check that list_groups is valid for group matrix construction.
+
+    - Is a list of list
+    - Does not contain the same feature in different groups
+    - Does not contain unknown features (>= input_dim)
+    - Does not contain empty groups.
+
+    Parameters
+    ----------
+    list_groups : list of list of int
+        Each element is a list representing features in the same group.
+        One feature should appear in maximum one group.
+        Feature that don't get assigned a group will be in their own group of one feature.
+    input_dim : int
+        Number of features in the initial dataset
+
+    """
+    assert isinstance(list_groups, list), "list_groups must be a list of list."
+
+    if len(list_groups) == 0:
+        return
+    else:
+        for group_pos, group in enumerate(list_groups):
+            msg = f"Groups must be given as a list of list, but found {group} in position {group_pos}."  # noqa
+            assert isinstance(group, list), msg
+            assert len(group) > 0, "Empty groups are forbidding please remove empty groups []"
+
+    n_elements_in_groups = np.sum([len(group) for group in list_groups])
+    flat_list = []
+    for group in list_groups:
+        flat_list.extend(group)
+    unique_elements = np.unique(flat_list)
+    n_unique_elements_in_groups = len(unique_elements)
+    msg = "One feature can only appear in one group, please check your grouped_features."
+    assert n_unique_elements_in_groups == n_elements_in_groups, msg
+
+    highest_feat = np.max(unique_elements)
+    assert highest_feat < input_dim, f"Number of features is {input_dim} but one group contains {highest_feat}."  # noqa
+    return
